+++ app/src/regexpEscape.js
@@ -1,0 +1,3 @@
+export default function regexpEscape(str) {
+  return str.replace(/[-[\]{}()*+?.,\\^$|#\s]/g, '\\$&');
+}